requirement.txt
requests
pandasgit 